+++ pkg/deployments/tasks/deprecated/20230206-composable-stable-pool-v3/index.ts
@@ -1,11 +1,11 @@
-import Task, { TaskMode } from '../../src/task';
-import { TaskRunOptions } from '../../src/types';
+import Task, { TaskMode } from '../../../src/task';
+import { TaskRunOptions } from '../../../src/types';
 import { ComposableStablePoolDeployment } from './input';
 
 import { ZERO_ADDRESS } from '@balancer-labs/v2-helpers/src/constants';
 import { bn } from '@balancer-labs/v2-helpers/src/numbers';
 import * as expectEvent from '@balancer-labs/v2-helpers/src/test/expectEvent';
-import { getContractDeploymentTransactionHash, saveContractDeploymentTransactionHash } from '../../src/network';
+import { getContractDeploymentTransactionHash, saveContractDeploymentTransactionHash } from '../../../src/network';
 import { ethers } from 'hardhat';
 
 export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {